+++ eck-autocomplete/index.ts
@@ -1,2 +1,0 @@
-export * from './eck-autocomplete-component/eck-autocomplete-component';
-export * from './eck-autocomplete-option-component/eck-autocomplete-option-component';