+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - By default, when no anchor is specified, the panel will continue to be attached to the connected input.
   - The anchor element can also be passed in programmatically when calling `setInputRef` as an optional second parameter.
 - fix: don't try to remove event listeners on disconnect if no input was connected
+- fix: export all types from main entry point and provide an exports field in package.json
 
 ## 1.1.3